Moved GameBoard class into models and integrated with Initializer
