+++ server/src/tests/models/InitializeScoreBoard.test.ts
@@ -1,8 +1,12 @@
 import { initializeScoreBoards } from "../../models/InitializeScoreBoard";
+import GameBoard from "../../models/GameBoardTemp";
 
 describe("initializeScoreBoards tests", () => {
   test("given 2 player names, it returns an array of 2 ScoreBoard objects", () => {
     const scoreBoardsArray = initializeScoreBoards(["player1", "player2"]);
+    scoreBoardsArray.forEach((scoreBoard) => {
+      expect(scoreBoard instanceof GameBoard).toBe(true); 
+    })
     expect(scoreBoardsArray.length).toBe(2);
   });
